[GEOPY-2414] Add scaling option. Refactor functions
